+++ app/src/main/java/com/qhy040404/libraryonetap/secret/SecretActivity.kt
@@ -2,8 +2,8 @@ package com.qhy040404.libraryonetap.secret
 
 import android.net.Uri
 import com.qhy040404.libraryonetap.R
-import com.qhy040404.libraryonetap.StartUpActivity
-import com.qhy040404.libraryonetap.views.ModifiedVideoView
+import com.qhy040404.libraryonetap.activity.StartUpActivity
+import com.qhy040404.libraryonetap.view.ModifiedVideoView
 
 class SecretActivity : StartUpActivity() {
     override fun init() = Thread(Play()).start()
