+++ app/src/main/res/layout/activity_secret.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".secret.SecretActivity">
 
-    <com.qhy040404.libraryonetap.views.ModifiedVideoView
+    <com.qhy040404.libraryonetap.view.ModifiedVideoView
         android:id="@+id/videoView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
