+++ app/src/main/res/layout/activity_yanxiujian.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".YanxiujianActivity">
+    tools:context=".activity.YanxiujianActivity">
 
     <TextView
         android:id="@+id/textView2"
